SyntaxHighlighter splitted up to small pieces
