+++ FastColoredTextBox/Models/Traits/HTML.cs
@@ -1,0 +1,123 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Globalization;
+using System.IO;
+using System.Text.RegularExpressions;
+using System.Xml;
+
+namespace FastColoredTextBoxNS
+{
+    public partial class SyntaxHighlighter : IDisposable
+    {
+        protected Regex HTMLAttrRegex,
+                      HTMLAttrValRegex,
+                      HTMLCommentRegex1,
+                      HTMLCommentRegex2;
+
+        protected Regex HTMLEndTagRegex;
+
+        protected Regex HTMLEntityRegex,
+                      HTMLTagContentRegex;
+
+        protected Regex HTMLTagNameRegex;
+        protected Regex HTMLTagRegex;
+
+        protected void HTMLAutoIndentNeeded(object sender, AutoIndentEventArgs args)
+        {
+            var tb = sender as FastColoredTextBox;
+            tb.CalcAutoIndentShiftByCodeFolding(sender, args);
+        }
+
+        protected void InitHTMLRegex()
+        {
+            HTMLCommentRegex1 = new Regex(@"(<!--.*?-->)|(<!--.*)", RegexOptions.Singleline | RegexCompiledOption);
+            HTMLCommentRegex2 = new Regex(@"(<!--.*?-->)|(.*-->)",
+                                          RegexOptions.Singleline | RegexOptions.RightToLeft | RegexCompiledOption);
+            HTMLTagRegex = new Regex(@"<|/>|</|>", RegexCompiledOption);
+            HTMLTagNameRegex = new Regex(@"<(?<range>[!\w:]+)", RegexCompiledOption);
+            HTMLEndTagRegex = new Regex(@"</(?<range>[\w:]+)>", RegexCompiledOption);
+            HTMLTagContentRegex = new Regex(@"<[^>]+>", RegexCompiledOption);
+            HTMLAttrRegex =
+                new Regex(
+                    @"(?<range>[\w\d\-]{1,20}?)='[^']*'|(?<range>[\w\d\-]{1,20})=""[^""]*""|(?<range>[\w\d\-]{1,20})=[\w\d\-]{1,20}",
+                    RegexCompiledOption);
+            HTMLAttrValRegex =
+                new Regex(
+                    @"[\w\d\-]{1,20}?=(?<range>'[^']*')|[\w\d\-]{1,20}=(?<range>""[^""]*"")|[\w\d\-]{1,20}=(?<range>[\w\d\-]{1,20})",
+                    RegexCompiledOption);
+            HTMLEntityRegex = new Regex(@"\&(amp|gt|lt|nbsp|quot|apos|copy|reg|#[0-9]{1,8}|#x[0-9a-f]{1,8});",
+                                        RegexCompiledOption | RegexOptions.IgnoreCase);
+        }
+
+        /// <summary>
+        /// Highlights HTML code
+        /// </summary>
+        /// <param name="range"></param>
+        public virtual void HTMLSyntaxHighlight(Range range)
+        {
+            range.tb.CommentPrefix = null;
+            range.tb.LeftBracket = '<';
+            range.tb.RightBracket = '>';
+            range.tb.LeftBracket2 = '(';
+            range.tb.RightBracket2 = ')';
+            range.tb.AutoIndentCharsPatterns = @"";
+            //clear style of changed range
+            range.ClearStyle(CommentStyle, TagBracketStyle, TagNameStyle, AttributeStyle, AttributeValueStyle,
+                             HtmlEntityStyle);
+            //
+            if (HTMLTagRegex == null)
+                InitHTMLRegex();
+            //comment highlighting
+            range.SetStyle(CommentStyle, HTMLCommentRegex1);
+            range.SetStyle(CommentStyle, HTMLCommentRegex2);
+            //tag brackets highlighting
+            range.SetStyle(TagBracketStyle, HTMLTagRegex);
+            //tag name
+            range.SetStyle(TagNameStyle, HTMLTagNameRegex);
+            //end of tag
+            range.SetStyle(TagNameStyle, HTMLEndTagRegex);
+            //attributes
+            range.SetStyle(AttributeStyle, HTMLAttrRegex);
+            //attribute values
+            range.SetStyle(AttributeValueStyle, HTMLAttrValRegex);
+            //html entity
+            range.SetStyle(HtmlEntityStyle, HTMLEntityRegex);
+
+            //clear folding markers
+            range.ClearFoldingMarkers();
+            //set folding markers
+            range.SetFoldingMarkers("<head", "</head>", RegexOptions.IgnoreCase);
+            range.SetFoldingMarkers("<body", "</body>", RegexOptions.IgnoreCase);
+            range.SetFoldingMarkers("<table", "</table>", RegexOptions.IgnoreCase);
+            range.SetFoldingMarkers("<form", "</form>", RegexOptions.IgnoreCase);
+            range.SetFoldingMarkers("<div", "</div>", RegexOptions.IgnoreCase);
+            range.SetFoldingMarkers("<script", "</script>", RegexOptions.IgnoreCase);
+            range.SetFoldingMarkers("<tr", "</tr>", RegexOptions.IgnoreCase);
+        }
+
+        #region Styles
+
+        /// <summary>
+        /// HTML Entity style
+        /// </summary>
+        public Style HtmlEntityStyle { get; set; }
+
+        /// <summary>
+        /// HTML attribute value style
+        /// </summary>
+        public Style AttributeValueStyle { get; set; }
+
+        /// <summary>
+        /// HTML tag brackets style
+        /// </summary>
+        public Style TagBracketStyle { get; set; }
+
+        /// <summary>
+        /// HTML tag name style
+        /// </summary>
+        public Style TagNameStyle { get; set; }
+
+        #endregion
+    }
+}
