+++ FastColoredTextBox/Models/Traits/JavaScript.cs
@@ -1,0 +1,76 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Globalization;
+using System.IO;
+using System.Text.RegularExpressions;
+using System.Xml;
+
+namespace FastColoredTextBoxNS
+{
+    public partial class SyntaxHighlighter : IDisposable
+    {
+        protected Regex JScriptCommentRegex1,
+                      JScriptCommentRegex2,
+                      JScriptCommentRegex3;
+
+        protected Regex JScriptKeywordRegex;
+        protected Regex JScriptNumberRegex;
+        protected Regex JScriptStringRegex;
+
+        protected void InitJScriptRegex()
+        {
+            JScriptStringRegex = new Regex(@"""""|''|"".*?[^\\]""|'.*?[^\\]'", RegexCompiledOption);
+            JScriptCommentRegex1 = new Regex(@"//.*$", RegexOptions.Multiline | RegexCompiledOption);
+            JScriptCommentRegex2 = new Regex(@"(/\*.*?\*/)|(/\*.*)", RegexOptions.Singleline | RegexCompiledOption);
+            JScriptCommentRegex3 = new Regex(@"(/\*.*?\*/)|(.*\*/)",
+                                             RegexOptions.Singleline | RegexOptions.RightToLeft | RegexCompiledOption);
+            JScriptNumberRegex = new Regex(@"\b\d+[\.]?\d*([eE]\-?\d+)?[lLdDfF]?\b|\b0x[a-fA-F\d]+\b",
+                                           RegexCompiledOption);
+            JScriptKeywordRegex =
+                new Regex(
+                    @"\b(true|false|break|case|catch|const|continue|default|delete|do|else|export|for|function|if|in|instanceof|new|null|return|switch|this|throw|try|var|void|while|with|typeof)\b",
+                    RegexCompiledOption);
+        }
+
+        /// <summary>
+        /// Highlights JavaScript code
+        /// </summary>
+        /// <param name="range"></param>
+        public virtual void JScriptSyntaxHighlight(Range range)
+        {
+            range.tb.CommentPrefix = "//";
+            range.tb.LeftBracket = '(';
+            range.tb.RightBracket = ')';
+            range.tb.LeftBracket2 = '{';
+            range.tb.RightBracket2 = '}';
+            range.tb.BracketsHighlightStrategy = BracketsHighlightStrategy.Strategy2;
+
+            range.tb.AutoIndentCharsPatterns
+                = @"
+^\s*[\w\.]+(\s\w+)?\s*(?<range>=)\s*(?<range>[^;]+);
+";
+
+            //clear style of changed range
+            range.ClearStyle(StringStyle, CommentStyle, NumberStyle, KeywordStyle);
+            //
+            if (JScriptStringRegex == null)
+                InitJScriptRegex();
+            //string highlighting
+            range.SetStyle(StringStyle, JScriptStringRegex);
+            //comment highlighting
+            range.SetStyle(CommentStyle, JScriptCommentRegex1);
+            range.SetStyle(CommentStyle, JScriptCommentRegex2);
+            range.SetStyle(CommentStyle, JScriptCommentRegex3);
+            //number highlighting
+            range.SetStyle(NumberStyle, JScriptNumberRegex);
+            //keyword highlighting
+            range.SetStyle(KeywordStyle, JScriptKeywordRegex);
+            //clear folding markers
+            range.ClearFoldingMarkers();
+            //set folding markers
+            range.SetFoldingMarkers("{", "}"); //allow to collapse brackets block
+            range.SetFoldingMarkers(@"/\*", @"\*/"); //allow to collapse comment block
+        }
+    }
+}
