+++ FastColoredTextBox/Models/Traits/Lua.cs
@@ -1,0 +1,106 @@
+﻿using System;
+using System.Text.RegularExpressions;
+
+namespace FastColoredTextBoxNS
+{
+    public partial class SyntaxHighlighter : IDisposable
+    {
+        protected Regex LuaCommentRegex1 { get; private set; }
+        protected Regex LuaCommentRegex2 { get; private set; }
+        protected Regex LuaCommentRegex3 { get; private set; }
+
+        protected Regex LuaKeywordRegex { get; private set; }
+        protected Regex LuaNumberRegex { get; private set; }
+        protected Regex LuaStringRegex { get; private set; }
+        protected Regex LuaFunctionsRegex { get; private set; }
+
+        protected void InitLuaRegex()
+        {
+            LuaStringRegex = new Regex(@"""""|''|"".*?[^\\]""|'.*?[^\\]'", RegexCompiledOption);
+            LuaCommentRegex1 = new Regex(@"--.*$", RegexOptions.Multiline | RegexCompiledOption);
+            LuaCommentRegex2 = new Regex(@"(--\[\[.*?\]\])|(--\[\[.*)", RegexOptions.Singleline | RegexCompiledOption);
+            LuaCommentRegex3 = new Regex(@"(--\[\[.*?\]\])|(.*\]\])",
+                                             RegexOptions.Singleline | RegexOptions.RightToLeft | RegexCompiledOption);
+            LuaNumberRegex = new Regex(@"\b\d+[\.]?\d*([eE]\-?\d+)?[lLdDfF]?\b|\b0x[a-fA-F\d]+\b",
+                                           RegexCompiledOption);
+            LuaKeywordRegex =
+                new Regex(
+                    @"\b(and|break|do|else|elseif|end|false|for|function|if|in|local|nil|not|or|repeat|return|then|true|until|while)\b",
+                    RegexCompiledOption);
+
+            LuaFunctionsRegex =
+                new Regex(
+                    @"\b(assert|collectgarbage|dofile|error|getfenv|getmetatable|ipairs|load|loadfile|loadstring|module|next|pairs|pcall|print|rawequal|rawget|rawset|require|select|setfenv|setmetatable|tonumber|tostring|type|unpack|xpcall)\b",
+                    RegexCompiledOption);
+        }
+
+        /// <summary>
+        /// Highlights Lua code
+        /// </summary>
+        /// <param name="range"></param>
+        public virtual void LuaSyntaxHighlight(Range range)
+        {
+            range.tb.CommentPrefix = "--";
+            range.tb.LeftBracket = '(';
+            range.tb.RightBracket = ')';
+            range.tb.LeftBracket2 = '{';
+            range.tb.RightBracket2 = '}';
+            range.tb.BracketsHighlightStrategy = BracketsHighlightStrategy.Strategy2;
+
+            range.tb.AutoIndentCharsPatterns
+                = @"
+^\s*[\w\.]+(\s\w+)?\s*(?<range>=)\s*(?<range>.+)
+";
+
+            //clear style of changed range
+            range.ClearStyle(StringStyle, CommentStyle, NumberStyle, KeywordStyle, FunctionsStyle);
+            //
+            if (LuaStringRegex == null)
+                InitLuaRegex();
+            //string highlighting
+            range.SetStyle(StringStyle, LuaStringRegex);
+            //comment highlighting
+            range.SetStyle(CommentStyle, LuaCommentRegex1);
+            range.SetStyle(CommentStyle, LuaCommentRegex2);
+            range.SetStyle(CommentStyle, LuaCommentRegex3);
+            //number highlighting
+            range.SetStyle(NumberStyle, LuaNumberRegex);
+            //keyword highlighting
+            range.SetStyle(KeywordStyle, LuaKeywordRegex);
+            //functions highlighting
+            range.SetStyle(FunctionsStyle, LuaFunctionsRegex);
+            //clear folding markers
+            range.ClearFoldingMarkers();
+            //set folding markers
+            range.SetFoldingMarkers("{", "}"); //allow to collapse brackets block
+            range.SetFoldingMarkers(@"--\[\[", @"\]\]"); //allow to collapse comment block
+        }
+
+        protected void LuaAutoIndentNeeded(object sender, AutoIndentEventArgs args)
+        {
+            //end of block
+            if (Regex.IsMatch(args.LineText, @"^\s*(end|until)\b"))
+            {
+                args.Shift = -args.TabLength;
+                args.ShiftNextLines = -args.TabLength;
+                return;
+            }
+            // then ...
+            if (Regex.IsMatch(args.LineText, @"\b(then)\s*\S+"))
+                return;
+            //start of operator block
+            if (Regex.IsMatch(args.LineText, @"^\s*(function|do|for|while|repeat|if)\b"))
+            {
+                args.ShiftNextLines = args.TabLength;
+                return;
+            }
+
+            //Statements else, elseif, case etc
+            if (Regex.IsMatch(args.LineText, @"^\s*(else|elseif)\b", RegexOptions.IgnoreCase))
+            {
+                args.Shift = -args.TabLength;
+                return;
+            }
+        }
+    }
+}
