+++ FastColoredTextBox/Models/Traits/XML.cs
@@ -1,0 +1,265 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Drawing;
+using System.Globalization;
+using System.IO;
+using System.Text.RegularExpressions;
+using System.Xml;
+
+namespace FastColoredTextBoxNS
+{
+    public partial class SyntaxHighlighter : IDisposable
+    {
+        protected Regex XMLAttrRegex,
+                      XMLAttrValRegex,
+                      XMLCommentRegex1,
+                      XMLCommentRegex2;
+
+        protected Regex XMLEndTagRegex;
+
+        protected Regex XMLEntityRegex,
+                      XMLTagContentRegex;
+
+        protected Regex XMLTagNameRegex;
+        protected Regex XMLTagRegex;
+        protected Regex XMLCDataRegex;
+        protected Regex XMLFoldingRegex;
+
+        protected void XMLAutoIndentNeeded(object sender, AutoIndentEventArgs args)
+        {
+            var tb = sender as FastColoredTextBox;
+            tb.CalcAutoIndentShiftByCodeFolding(sender, args);
+        }
+
+        /// <summary>
+        /// Uses the given <paramref name="doc"/> to parse a XML description and adds it as syntax descriptor. 
+        /// The syntax descriptor is used for highlighting when 
+        /// <list type="bullet">
+        ///     <item>Language property of FCTB is set to <see cref="Language.Custom"/></item>
+        ///     <item>DescriptionFile property of FCTB has the same value as the method parameter <paramref name="descriptionFileName"/></item>
+        /// </list>
+        /// </summary>
+        /// <param name="descriptionFileName">Name of the description file</param>
+        /// <param name="doc">XmlDocument to parse</param>
+        public virtual void AddXmlDescription(string descriptionFileName, XmlDocument doc)
+        {
+            SyntaxDescriptor desc = ParseXmlDescription(doc);
+            descByXMLfileNames[descriptionFileName] = desc;
+        }
+
+        public static SyntaxDescriptor ParseXmlDescription(XmlDocument doc)
+        {
+            var desc = new SyntaxDescriptor();
+            XmlNode brackets = doc.SelectSingleNode("doc/brackets");
+            if (brackets != null)
+            {
+                if (brackets.Attributes["left"] == null || brackets.Attributes["right"] == null ||
+                    brackets.Attributes["left"].Value == "" || brackets.Attributes["right"].Value == "")
+                {
+                    desc.leftBracket = '\x0';
+                    desc.rightBracket = '\x0';
+                }
+                else
+                {
+                    desc.leftBracket = brackets.Attributes["left"].Value[0];
+                    desc.rightBracket = brackets.Attributes["right"].Value[0];
+                }
+
+                if (brackets.Attributes["left2"] == null || brackets.Attributes["right2"] == null ||
+                    brackets.Attributes["left2"].Value == "" || brackets.Attributes["right2"].Value == "")
+                {
+                    desc.leftBracket2 = '\x0';
+                    desc.rightBracket2 = '\x0';
+                }
+                else
+                {
+                    desc.leftBracket2 = brackets.Attributes["left2"].Value[0];
+                    desc.rightBracket2 = brackets.Attributes["right2"].Value[0];
+                }
+
+                if (brackets.Attributes["strategy"] == null || brackets.Attributes["strategy"].Value == "")
+                    desc.bracketsHighlightStrategy = BracketsHighlightStrategy.Strategy2;
+                else
+                    desc.bracketsHighlightStrategy = (BracketsHighlightStrategy)Enum.Parse(typeof(BracketsHighlightStrategy), brackets.Attributes["strategy"].Value);
+            }
+
+            var styleByName = new Dictionary<string, Style>();
+
+            foreach (XmlNode style in doc.SelectNodes("doc/style"))
+            {
+                Style s = ParseStyle(style);
+                styleByName[style.Attributes["name"].Value] = s;
+                desc.styles.Add(s);
+            }
+            foreach (XmlNode rule in doc.SelectNodes("doc/rule"))
+                desc.rules.Add(ParseRule(rule, styleByName));
+            foreach (XmlNode folding in doc.SelectNodes("doc/folding"))
+                desc.foldings.Add(ParseFolding(folding));
+
+            return desc;
+        }
+
+        protected void InitXMLRegex()
+        {
+            XMLCommentRegex1 = new Regex(@"(<!--.*?-->)|(<!--.*)", RegexOptions.Singleline | RegexCompiledOption);
+            XMLCommentRegex2 = new Regex(@"(<!--.*?-->)|(.*-->)",
+                                          RegexOptions.Singleline | RegexOptions.RightToLeft | RegexCompiledOption);
+            XMLTagRegex = new Regex(@"<\?|<|/>|</|>|\?>", RegexCompiledOption);
+            XMLTagNameRegex = new Regex(@"<[?](?<range1>[x][m][l]{1})|<(?<range>[!\w:]+)", RegexCompiledOption);
+            XMLEndTagRegex = new Regex(@"</(?<range>[\w:]+)>", RegexCompiledOption);
+            XMLTagContentRegex = new Regex(@"<[^>]+>", RegexCompiledOption);
+            XMLAttrRegex =
+                new Regex(
+                    @"(?<range>[\w\d\-\:]+)[ ]*=[ ]*'[^']*'|(?<range>[\w\d\-\:]+)[ ]*=[ ]*""[^""]*""|(?<range>[\w\d\-\:]+)[ ]*=[ ]*[\w\d\-\:]+",
+                    RegexCompiledOption);
+            XMLAttrValRegex =
+                new Regex(
+                    @"[\w\d\-]+?=(?<range>'[^']*')|[\w\d\-]+[ ]*=[ ]*(?<range>""[^""]*"")|[\w\d\-]+[ ]*=[ ]*(?<range>[\w\d\-]+)",
+                    RegexCompiledOption);
+            XMLEntityRegex = new Regex(@"\&(amp|gt|lt|nbsp|quot|apos|copy|reg|#[0-9]{1,8}|#x[0-9a-f]{1,8});",
+                                        RegexCompiledOption | RegexOptions.IgnoreCase);
+            XMLCDataRegex = new Regex(@"<!\s*\[CDATA\s*\[(?<text>(?>[^]]+|](?!]>))*)]]>", RegexCompiledOption | RegexOptions.IgnoreCase); // http://stackoverflow.com/questions/21681861/i-need-a-regex-that-matches-cdata-elements-in-html
+            XMLFoldingRegex = new Regex(@"<(?<range>/?\w+)\s[^>]*?[^/]>|<(?<range>/?\w+)\s*>", RegexOptions.Singleline | RegexCompiledOption);
+        }
+
+        /// <summary>
+        /// Highlights XML code
+        /// </summary>
+        /// <param name="range"></param>
+        public virtual void XMLSyntaxHighlight(Range range)
+        {
+            range.tb.CommentPrefix = null;
+            range.tb.LeftBracket = '<';
+            range.tb.RightBracket = '>';
+            range.tb.LeftBracket2 = '(';
+            range.tb.RightBracket2 = ')';
+            range.tb.AutoIndentCharsPatterns = @"";
+            //clear style of changed range
+            range.ClearStyle(CommentStyle, XmlTagBracketStyle, XmlTagNameStyle, XmlAttributeStyle, XmlAttributeValueStyle,
+                             XmlEntityStyle, XmlCDataStyle);
+
+            //
+            if (XMLTagRegex == null)
+            {
+                InitXMLRegex();
+            }
+
+            //xml CData
+            range.SetStyle(XmlCDataStyle, XMLCDataRegex);
+
+            //comment highlighting
+            range.SetStyle(CommentStyle, XMLCommentRegex1);
+            range.SetStyle(CommentStyle, XMLCommentRegex2);
+
+            //tag brackets highlighting
+            range.SetStyle(XmlTagBracketStyle, XMLTagRegex);
+
+            //tag name
+            range.SetStyle(XmlTagNameStyle, XMLTagNameRegex);
+
+            //end of tag
+            range.SetStyle(XmlTagNameStyle, XMLEndTagRegex);
+
+            //attributes
+            range.SetStyle(XmlAttributeStyle, XMLAttrRegex);
+
+            //attribute values
+            range.SetStyle(XmlAttributeValueStyle, XMLAttrValRegex);
+
+            //xml entity
+            range.SetStyle(XmlEntityStyle, XMLEntityRegex);
+
+            //clear folding markers
+            range.ClearFoldingMarkers();
+
+            //set folding markers
+            XmlFolding(range);
+        }
+
+        private void XmlFolding(Range range)
+        {
+            var stack = new Stack<XmlFoldingTag>();
+            var id = 0;
+            var fctb = range.tb;
+            //extract opening and closing tags (exclude open-close tags: <TAG/>)
+            foreach (var r in range.GetRanges(XMLFoldingRegex))
+            {
+                var tagName = r.Text;
+                var iLine = r.Start.iLine;
+                //if it is opening tag...
+                if (tagName[0] != '/')
+                {
+                    // ...push into stack
+                    var tag = new XmlFoldingTag { Name = tagName, id = id++, startLine = r.Start.iLine };
+                    stack.Push(tag);
+                    // if this line has no markers - set marker
+                    if (string.IsNullOrEmpty(fctb[iLine].FoldingStartMarker))
+                        fctb[iLine].FoldingStartMarker = tag.Marker;
+                }
+                else
+                {
+                    //if it is closing tag - pop from stack
+                    if (stack.Count > 0)
+                    {
+                        var tag = stack.Pop();
+                        //compare line number
+                        if (iLine == tag.startLine)
+                        {
+                            //remove marker, because same line can not be folding
+                            if (fctb[iLine].FoldingStartMarker == tag.Marker) //was it our marker?
+                                fctb[iLine].FoldingStartMarker = null;
+                        }
+                        else
+                        {
+                            //set end folding marker
+                            if (string.IsNullOrEmpty(fctb[iLine].FoldingEndMarker))
+                                fctb[iLine].FoldingEndMarker = tag.Marker;
+                        }
+                    }
+                }
+            }
+        }
+
+        class XmlFoldingTag
+        {
+            public string Name;
+            public int id;
+            public int startLine;
+            public string Marker { get { return Name + id; } }
+        }
+
+        #region Styles
+        
+        /// <summary>
+        /// XML attribute style
+        /// </summary>
+        public Style XmlAttributeStyle { get; set; }
+
+        /// <summary>
+        /// XML attribute value style
+        /// </summary>
+        public Style XmlAttributeValueStyle { get; set; }
+
+        /// <summary>
+        /// XML tag brackets style
+        /// </summary>
+        public Style XmlTagBracketStyle { get; set; }
+
+        /// <summary>
+        /// XML tag name style
+        /// </summary>
+        public Style XmlTagNameStyle { get; set; }
+
+        /// <summary>
+        /// XML Entity style
+        /// </summary>
+        public Style XmlEntityStyle { get; set; }
+
+        /// <summary>
+        /// XML CData style
+        /// </summary>
+        public Style XmlCDataStyle { get; set; }
+
+        #endregion
+    }
+}
